+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to the Pony compiler and standard library will be documented
 
 ### Changed
 
-- Remove `glob` package from standard library ([PR #3219](https://github.com/ponylang/ponyc/pull/3219))
+- Remove `glob` package from standard library ([PR #3220](https://github.com/ponylang/ponyc/pull/3220))
 
 ## [0.29.0] - 2019-07-06