+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to the Pony compiler and standard library will be documented
 
 ### Changed
 
+- Stop building Tumbleweed packages for releases ([PR #3228](https://github.com/ponylang/ponyc/pull/3228))
 - Remove `glob` package from standard library ([PR #3220](https://github.com/ponylang/ponyc/pull/3220))
 - Remove `regex` package from standard library ([PR #3218](https://github.com/ponylang/ponyc/pull/3218))