+++ CHANGELOG.md
@@ -16,6 +16,8 @@ All notable changes to the Pony compiler and standard library will be documented
 - Stop creating Debian Buster releases ([PR #3227](https://github.com/ponylang/ponyc/pull/3227))
 - Remove `glob` package from standard library ([PR #3220](https://github.com/ponylang/ponyc/pull/3220))
 - Remove `regex` package from standard library ([PR #3218](https://github.com/ponylang/ponyc/pull/3218))
+- Remove `crypto` package from standard library ([PR #3225](https://github.com/ponylang/ponyc/pull/3225))
+- Remove `net/ssl` package from standard library ([PR #3225](https://github.com/ponylang/ponyc/pull/3225))
 
 ## [0.29.0] - 2019-07-06